Added includeScript option and test fixes
Added an includeScript flag for the configuration recaptcha { .. } block
which, if false, does not output script tags to the response (similar to
includeNoScript).  It defaults to true to maintain backwards
compatibility.

Also, this commit fixes broken unit test cases and updates them for the
v2.0 API. Also added a test for the new flag.